chore: Tidy up code
