+++ src/ExternalSearch.Providers.Gleif/ExternalSearch.Providers.Gleif.csproj
@@ -5,13 +5,7 @@
   <ItemGroup>
     <EmbeddedResource Include="Resources\gleif.png" />
   </ItemGroup>
-    <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-        <LangVersion>latest</LangVersion>
-    </PropertyGroup>
 
-    <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-        <LangVersion>latest</LangVersion>
-    </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="CluedIn.Core" />
     <PackageReference Include="CluedIn.ExternalSearch" />
