+++ src/ExternalSearch.Providers.Gleif/Models/Attributes.cs
@@ -1,0 +1,15 @@
+using Newtonsoft.Json;
+
+namespace CluedIn.ExternalSearch.Providers.Gleif.Models;
+
+public class Attributes
+{
+    [JsonProperty("lei")]
+    public string Lei { get; set; }
+
+    [JsonProperty("entity")]
+    public Entity Entity { get; set; }
+
+    [JsonProperty("registration")]
+    public Registration Registration { get; set; }
+}
