+++ src/ExternalSearch.Providers.Gleif/Models/Data.cs
@@ -1,0 +1,15 @@
+using Newtonsoft.Json;
+
+namespace CluedIn.ExternalSearch.Providers.Gleif.Models;
+
+public class Data
+{
+    [JsonProperty("type")]
+    public string Type { get; set; }
+
+    [JsonProperty("id")]
+    public string Id { get; set; }
+
+    [JsonProperty("attributes")]
+    public Attributes Attributes { get; set; }
+}
