--- src/ExternalSearch.Providers.Gleif/Models/DataValue.cs
@@ -1,10 +1,0 @@
-using Newtonsoft.Json;
-
-namespace CluedIn.ExternalSearch.Providers.Gleif.Models
-{
-    //public class DataValue
-    //{
-    //    [JsonProperty("$")]
-    //    public string Value { get; set; }
-    //}
-}
