+++ src/ExternalSearch.Providers.Gleif/Models/Expiration.cs
@@ -1,0 +1,12 @@
+using Newtonsoft.Json;
+
+namespace CluedIn.ExternalSearch.Providers.Gleif.Models;
+
+public class Expiration
+{
+    [JsonProperty("date")]
+    public object Date { get; set; }
+
+    [JsonProperty("reason")]
+    public object Reason { get; set; }
+}
