+++ src/ExternalSearch.Providers.Gleif/Models/LegalName.cs
@@ -1,0 +1,12 @@
+using Newtonsoft.Json;
+
+namespace CluedIn.ExternalSearch.Providers.Gleif.Models;
+
+public class LegalName
+{
+    [JsonProperty("name")]
+    public string Name { get; set; }
+
+    [JsonProperty("language")]
+    public string Language { get; set; }
+}
