+++ src/ExternalSearch.Providers.Gleif/Models/Metadata.cs
@@ -1,0 +1,5 @@
+namespace CluedIn.ExternalSearch.Providers.Gleif.Models;
+
+public class Metadata
+{
+}
