--- src/ExternalSearch.Providers.Gleif/Models/OtherAddress.cs
@@ -1,11 +1,0 @@
-using System.Collections.Generic;
-using Newtonsoft.Json;
-
-namespace CluedIn.ExternalSearch.Providers.Gleif.Models
-{
-    //public class OtherAddress
-    //{
-    //    [JsonProperty("OtherAddress")]
-    //    public List<Address> Addresses { get; set; }
-    //}
-}
