--- src/ExternalSearch.Providers.Gleif/Models/OtherEntityName.cs
@@ -1,11 +1,0 @@
-using System.Collections.Generic;
-using Newtonsoft.Json;
-
-namespace CluedIn.ExternalSearch.Providers.Gleif.Models
-{
-    //public class OtherEntityName
-    //{
-    //    [JsonProperty("OtherEntityName")]
-    //    public List<DataValue> Names { get; set; }
-    //}
-}
