+++ src/ExternalSearch.Providers.Gleif/Models/OtherName.cs
@@ -1,0 +1,15 @@
+using Newtonsoft.Json;
+
+namespace CluedIn.ExternalSearch.Providers.Gleif.Models;
+
+public class OtherName
+{
+    [JsonProperty("name")]
+    public string Name { get; set; }
+
+    [JsonProperty("language")]
+    public string Language { get; set; }
+
+    [JsonProperty("type")]
+    public string Type { get; set; }
+}
