--- src/ExternalSearch.Providers.Gleif/Models/RegistrationAuthority.cs
@@ -1,16 +1,0 @@
-using Newtonsoft.Json;
-
-namespace CluedIn.ExternalSearch.Providers.Gleif.Models
-{
-    //public class RegistrationAuthority
-    //{
-    //    [JsonProperty("RegistrationAuthorityID")]
-    //    public DataValue RegistrationAuthorityId { get; set; }
-
-    //    [JsonProperty("OtherRegistrationAuthorityID")]
-    //    public DataValue OtherRegistrationAuthorityId { get; set; }
-
-    //    [JsonProperty("RegistrationAuthorityEntityID")]
-    //    public DataValue RegistrationAuthorityEntityId { get; set; }
-    //}
-}
