--- src/ExternalSearch.Providers.Gleif/Models/TransliteratedOtherAddress.cs
@@ -1,11 +1,0 @@
-using System.Collections.Generic;
-using Newtonsoft.Json;
-
-namespace CluedIn.ExternalSearch.Providers.Gleif.Models
-{
-    //public class TransliteratedOtherAddress
-    //{
-    //    [JsonProperty("TransliteratedOtherAddress")]
-    //    public List<Address> Addresses { get; set; }
-    //}
-}
