+++ src/ExternalSearch.Providers.Gleif/Models/ValidatedAt.cs
@@ -2,7 +2,7 @@ using Newtonsoft.Json;
 
 namespace CluedIn.ExternalSearch.Providers.Gleif.Models;
 
-public class LegalForm
+public class ValidatedAt
 {
     [JsonProperty("id")]
     public string Id { get; set; }
